restructure idtype
