+++ src/idtype/ProductIDType.ts
@@ -1,0 +1,254 @@
+/* *****************************************************************************
+ * Caleydo - Visualization for Molecular Biology - http://caleydo.org
+ * Copyright (c) The Caleydo Team. All rights reserved.
+ * Licensed under the new BSD license, available at http://caleydo.org/license
+ **************************************************************************** */
+/**
+ * Created by Samuel Gratzl on 04.08.2014.
+ */
+
+import {EventHandler, IEvent, IEventListener} from '../event';
+import {none, Range, RangeLike, parse, Range1D, list as rlist} from '../range';
+import {SelectOperation, asSelectOperation, IIDType, defaultSelectionType} from './IIDType';
+import IDType from './IDType';
+
+//function indicesCompare(a: number[], b: number[]) {
+//  //assert a.length = b.length
+//  for(let i = 0; i < a.length; ++i) {
+//    if (a[i] !== b[i]) {
+//      return a[i] - b[i];
+//    }
+//  }
+//  return 0;
+//}
+//
+//function compressPairs(pairs: number[][]): Range[] {
+//  return pairs.map((a) => rlist(...a));
+//}
+
+function overlaps(r: Range, with_: Range, ndim: number) {
+  if (with_.ndim === 0) {
+    return true; //catch all
+  }
+  for (let i = 0; i < Math.min(r.ndim, ndim); ++i) {
+    let ri = r.dim(i);
+    let wi = with_.dim(i);
+    if (wi.isAll || ri.isAll) {
+      return true;
+    }
+    if (!ri.isUnbound && ri.asList().every((rii) => !wi.contains(rii))) {
+      //it the ids at dimension i are not overlapping can't overlap in others
+      return false;
+    }
+    //TODO
+  }
+  return false;
+}
+
+function removeCells(b: Range[], without: Range[], ndim: number) {
+  if (without.length === 0) {
+    return b;
+  }
+  let r: Range[] = [];
+  b.forEach((bi) => {
+    if (without.some((w) => w.eq(bi))) {
+      //skip
+    } else if (without.some((w) => overlaps(bi, w, ndim))) {
+      //TODO
+    } else {
+      r.push(bi);
+    }
+  });
+  return r;
+}
+/**
+ * a product idtype is a product of multiple underlying ones, e.g. patient x gene.
+ */
+export default class ProductIDType extends EventHandler implements IIDType {
+  static readonly EVENT_SELECT_DIM = 'selectDim';
+  static readonly EVENT_SELECT_PRODUCT = 'selectProduct';
+
+  private sel = new Map<string, Range[]>();
+
+  private isOn = false;
+  private selectionListener = (event: IEvent, type: string, act: Range, added: Range, removed: Range) => {
+    this.fire(`${ProductIDType.EVENT_SELECT_DIM},${ProductIDType.EVENT_SELECT_PRODUCT}`, this.elems.indexOf(<IDType>event.currentTarget), type, act, added, removed);
+    this.fire(`${ProductIDType.EVENT_SELECT_DIM}-${type},${ProductIDType.EVENT_SELECT_PRODUCT}-${type}`, this.elems.indexOf(<IDType>event.currentTarget), act, added, removed);
+  };
+
+  constructor(public readonly elems: IDType[], public readonly internal = false) {
+    super();
+  }
+
+  on(events: string|{[key: string]: IEventListener}, listener?: IEventListener) {
+    if (!this.isOn) {
+      this.enable();
+      this.isOn = true;
+    }
+    return super.on(events, listener);
+  }
+
+  get id() {
+    return this.elems.map((e) => e.id).join('X');
+  }
+
+  get name() {
+    return this.elems.map((e) => e.name).join(' x ');
+  }
+
+  get names() {
+    return this.elems.map((e) => e.names).join(' x ');
+  }
+
+  private enable() {
+    this.elems.forEach((elem) => elem.on(IDType.EVENT_SELECT, this.selectionListener));
+  }
+
+  private disable() {
+    this.elems.forEach((elem) => elem.off(IDType.EVENT_SELECT, this.selectionListener));
+  }
+
+  persist() {
+    let s = {};
+    this.sel.forEach((v, type) => s[type] = v.map((r) => r.toString()));
+    return {
+      sel: s
+    };
+  }
+
+  restore(persisted: any) {
+    Object.keys(persisted.sel).forEach((type) => this.sel.set(type, persisted.sel[type].map(parse)));
+    return this;
+  }
+
+  toString() {
+    return this.name;
+  }
+
+  selectionTypes() {
+    return Array.from(this.sel.keys());
+  }
+
+  /**
+   * return the current selections of the given type
+   * @param type optional the selection type
+   * @returns {Range[]}
+   */
+  selections(type = defaultSelectionType): Range[] {
+    if (this.sel.has(type)) {
+      return this.sel.get(type).slice();
+    }
+    this.sel.set(type, []);
+    return [];
+  }
+
+  productSelections(type = defaultSelectionType /*, wildcardLookup: (idtype: IDType) => Promise<number> */): Range[] {
+    const cells = this.selections(type);
+    const usedCells = this.toPerDim(cells);
+    this.elems.forEach((e, i) => {
+      const s = e.selections(type);
+      //remove all already used rows / columns as part of the cells
+      const wildcard = s.without(usedCells[i]);
+      if (!wildcard.isNone) {
+        //create wildcard cells, e.g., the remaining ones are row/column selections
+        cells.push(rlist(this.elems.map((e2) => e === e2 ? wildcard.dim(0) : Range1D.all())));
+      }
+    });
+
+    return cells;
+    /* TODO no duplicates
+     if (cells.every((c) => !c.isUnbound)) {
+     //all cells are bound, just cells
+     return Promise.resolve(cells);
+     }
+     //we need to resolve some wildcards
+     return Promise.all(this.elems.map((elem, i) => {
+     if (cells.some((c) => c.dim(i).isUnbound)) {
+     return wildcardLookup(elem);
+     } else {
+     return Promise.resolve(0);
+     }
+     })).then((size: number[]) => {
+     const fullCells : any = {};
+     cells.forEach((cell) => {
+     cell.product((indices: number[]) => {
+     const id = indices.join('_');
+     fullCells[id] = indices;
+     });
+     }, size);
+     //fullCells contains all cells that we have to take care of
+     const pairs = Object.keys(fullCells).map((k) => fullCells[k]).sort(indicesCompare);
+     return compressPairs(pairs);
+     });
+     */
+  }
+
+  /**
+   * select the given range as
+   * @param range
+   */
+  select(range: RangeLike[]);
+  select(range: RangeLike[], op: SelectOperation);
+  select(type: string, range: RangeLike[]);
+  select(type: string, range: RangeLike[], op: SelectOperation);
+  select() {
+    const a = Array.from(arguments);
+    const type = (typeof a[0] === 'string') ? a.shift() : defaultSelectionType,
+      range = a[0].map(parse),
+      op = asSelectOperation(a[1]);
+    return this.selectImpl(range, op, type);
+  }
+
+  private selectImpl(cells: Range[], op = SelectOperation.SET, type: string = defaultSelectionType) {
+    const rcells = cells.map(parse);
+
+    const b = this.selections(type);
+
+    let new_: Range[] = [];
+
+    switch (op) {
+      case SelectOperation.SET:
+        new_ = rcells;
+        break;
+      case SelectOperation.ADD:
+        new_ = b.concat(rcells);
+        break;
+      case SelectOperation.REMOVE:
+        new_ = removeCells(b, rcells, this.elems.length);
+        break;
+    }
+    //if (b.eq(new_)) {
+    //  return b;
+    //}
+    this.sel[type] = new_;
+
+    //individual selection per dimension
+    const perDimSelections = this.toPerDim(new_);
+    this.disable();
+    this.elems.forEach((e, i) => e.select(type, perDimSelections[i]));
+    this.enable();
+
+    const added = op !== SelectOperation.REMOVE ? rcells : [];
+    const removed = (op === SelectOperation.ADD ? [] : (op === SelectOperation.SET ? b : rcells));
+    this.fire(IDType.EVENT_SELECT, type, new_, added, removed, b);
+    this.fire(ProductIDType.EVENT_SELECT_PRODUCT, -1, type, new_, added, removed, b);
+    this.fire(`${IDType.EVENT_SELECT}-${type}`, new_, added, removed, b);
+    this.fire(`${ProductIDType.EVENT_SELECT_PRODUCT}-${type}`, -1, new_, added, removed, b);
+    return b;
+  }
+
+  private toPerDim(sel: Range[]) {
+    return this.elems.map((elem, i) => {
+      if (sel.length === 0) {
+        return none();
+      }
+      const dimselections = sel.map((r) => r.dim(i));
+      const selection = dimselections.reduce((p, a) => p ? p.union(a) : a, null);
+      return rlist(selection);
+    });
+  }
+
+  clear(type = defaultSelectionType) {
+    return this.selectImpl([], SelectOperation.SET, type);
+  }
+}
